+++ rodi/dependency.py
@@ -1,0 +1,6 @@
+class Dependency:
+    __slots__ = ("name", "annotation")
+
+    def __init__(self, name, annotation):
+        self.name = name
+        self.annotation = annotation
