+++ rodi/exceptions/__init__.py
@@ -1,0 +1,20 @@
+from rodi.exceptions.di_exception import DIException
+from rodi.exceptions.alias_already_defined import AliasAlreadyDefined
+from rodi.exceptions.missing_type_exception import MissingTypeException
+from rodi.exceptions.circular_dependency_exception import CircularDependencyException
+from rodi.exceptions.alias_configuration_error import AliasConfigurationError
+from rodi.exceptions.cannot_resolve_parameter_exception import (
+    CannotResolveParameterException,
+)
+from rodi.exceptions.cannot_resolve_type_exception import CannotResolveTypeException
+from rodi.exceptions.factory_missing_context_exception import (
+    FactoryMissingContextException,
+)
+from rodi.exceptions.invalid_factory import InvalidFactory
+from rodi.exceptions.invalid_operation_in_strict_mode import (
+    InvalidOperationInStrictMode,
+)
+from rodi.exceptions.overriding_service_exception import OverridingServiceException
+from rodi.exceptions.unsupported_union_type_exception import (
+    UnsupportedUnionTypeException,
+)
