+++ rodi/exceptions/di_exception.py
@@ -1,0 +1,2 @@
+class DIException(Exception):
+    """Base exception class for DI exceptions."""
