+++ rodi/factory_wrappers/factory_wrapper_no_args.py
@@ -1,0 +1,8 @@
+class FactoryWrapperNoArgs:
+    __slots__ = ("factory",)
+
+    def __init__(self, factory):
+        self.factory = factory
+
+    def __call__(self, context, activating_type):
+        return self.factory()
