+++ rodi/service_life_style.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class ServiceLifeStyle(Enum):
+    TRANSIENT = 1
+    SCOPED = 2
+    SINGLETON = 3
