+++ rodi/utils/get_obj_locals.py
@@ -1,0 +1,5 @@
+from typing import Optional, Dict, Any
+
+
+def _get_obj_locals(obj) -> Optional[Dict[str, Any]]:
+    return getattr(obj, "_locals", None)
