+++ rodi/utils/get_plain_class_factory.py
@@ -1,0 +1,8 @@
+from typing import Type
+
+
+def _get_plain_class_factory(concrete_type: Type):
+    def factory(*args):
+        return concrete_type()
+
+    return factory
